fix: fixed utils
