--- src/resources/filters/deleteBoard.js
@@ -1,8 +1,0 @@
-const taskRepo = require('../../common/localDb');
-
-const deleteBoardsTask = (id) => {
-  taskRepo.TASKS = taskRepo.TASKS.filter((el) => id !== el.boardId);
-  return taskRepo.TASKS;
-};
-
-module.exports = deleteBoardsTask;
